+++ packages/core/src/types/curi.ts
@@ -48,7 +48,7 @@ export interface CurrentResponse {
 }
 
 export interface CuriRouter {
-  refresh: (routeArray: Array<RouteDescriptor>) => void;
+  replaceRoutes: (routeArray: Array<RouteDescriptor>) => void;
   respond: (
     fn: ResponseHandler,
     options?: RespondOptions