+++ packages/core/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Rename `router.refresh` to `router.replaceRoutes`.
+
 ## 1.0.0-beta.27
 
 * If no route has a `match.initial` or `match.every` function, the router will emit in synchronous mode. In synchronous mode, `router.respond` is not necessary for delaying the initial render.